pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jackal-paw"
version = "0.1.0"
description = "Plane-wave DFT code in Python using JAX"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "jax>=0.4.30",
  "jaxlib>=0.4.30",
  "numpy>=1.26",
  "scipy>=1.11",
  "ase>=3.23",
  "pyyaml>=6.0",
  "pydantic>=2.7",
  "lxml>=5.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "mypy>=1.10",
  "pre-commit>=3.7",
  "mkdocs>=1.6",
  "mkdocs-material>=9.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
